Make st.write use st.html when _repr_html_ is present.
This also changes some behavior:

1. When unsafe_allow_html=False AND _repr_html_() did not contain HTML tags
we used to send its output to st.markdown(). This behavior is now removed.
Instead, the output is handled like a normal object (that is, with st.help).

2. Removes _repr_html_ from *Connection objects. This didn't really buy us
anything, and was just weird. Also, this was the reason for adding behavior #1
in the first place.